Changed Rotation/Yaw to be Degrees everywhere instead of bytes or Quaternions in different parts. Fixed up mapLinks and moved around some NPCs

+++ AISpace.Common.Tests/SharedStateTests.cs
@@ -93,7 +93,7 @@ public class SharedStateTests
         public float X { get; set; }
         public float Y { get; set; }
         public float Z { get; set; }
-        public sbyte Rotation { get; set; }
+        public int Rotation { get; set; }
         public int MovementTypeId { get; set; }
         public bool HasMovedSinceMapLoad { get; set; }
         public bool IsMapTransitionPending { get; set; }
