+++ AISpace.Common/DAL/Entities/SessionPresence.cs
@@ -13,6 +13,6 @@ public class SessionPresence
     public float X { get; set; }
     public float Y { get; set; }
     public float Z { get; set; }
-    public sbyte Rotation { get; set; }
+    public int Rotation { get; set; }
     public DateTime UpdatedAtUtc { get; set; }
 }
